0008 - Fixed "lexicon.community" typo
The document references the group as `lexicons.community`, i.e. "lexicon" ending with an "s", but the actual group/domain is "lexicon" without the "s". <https://lexicon.community>

I wasn't sure if it was desirable to change the mention to a link. Either way, it seemed important to to get their name right... I still feel embarassed making a 1 character PR. :sweat_smile: